+++ international/en.php
@@ -639,7 +639,8 @@ $languages['en'] = array (
 	"label_opmlimporter" => "OPML Importer",
 	"config_playerintitlebar" => "Show Player Name in Titlebar",
 	"label_communityradio" => "Community Radio Browser",
-	"label_volume" => "Volume"
+	"label_volume" => "Volume",
+	"config_playlistswipe" => "Swipe to remove tracks from Current Playlist"
 );
 
 ?>