+++ radios/code/spotiCrazyRadio.js
@@ -17,7 +17,7 @@ var spotiCrazyRadio = function() {
 		populate: function(p, numtracks) {
 			if (typeof(spotifyRecommendationsRadio) == 'undefined') {
 				debug.log("CRAZY RADIO","Loading Spotify Radio Tuner");
-				$.getScript('radios/code/spotifyrecommendationsradio.js',function() {
+				$.getScript('radios/code/spotifyrecommendationsradio.js?version='+rompr_version,function() {
 					spotiCrazyRadio.actuallyGo(numtracks)
 				});
 			} else {